refactor(vendo): wire cohesion — context + doctor modules, review nits (kill-list B4)
